requirements.txt
asgiref==3.5.0
colorama==0.4.4
cycler==0.11.0
distlib==0.3.4
Django==4.0.2
django-admin-tools==0.9.2
django-crispy-forms==1.14.0
filelock==3.4.2
fonttools==4.28.2
future==0.18.2
kiwisolver==1.3.2
mando==0.6.4
matplotlib==3.5.0
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
platformdirs==2.4.1
pyparsing==3.0.6
python-dateutil==2.8.2
radon==5.1.0
setuptools-scm==6.3.2
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
tzdata==2021.5
virtualenv==20.13.1
virtualenvwrapper-win==1.2.7
